Homepage
Added homepage html that contains the basics of the configuration page for the configuration of the student who is registering. There is an option to select a major (A-Z) or undeclared at the top, and for students who are in a minor, they can select the checkbox which will cause a new scrolling menu of minors, and then at the bottom is organizations at Adelphi that the student can select that they are interested in. If they want to select multiple they can by checking the box and holding CTRL to then select multiple. However, this function is not fully working yet and is still in the process of fixing, as it still only shows one regardless. The user can enter their credentials and username at the top to register their values into the database and then when they click the 'submit' button at the bottom, a function will allow all of the necessary details to get put into SQL. 

I have also attached the jquery files as it will be used now and in the future for the website. The two images are the images used in the website homepage as well.